refactor(#2): move files from utils to util
