+++ src/push-notifications/send.ts
@@ -4,8 +4,8 @@ import { ServiceConfigurationOptions } from "aws-sdk/lib/service";
 import { IsInt, Max, Min, validate } from "class-validator";
 import Expo from "expo-server-sdk";
 import { InvalidArgument } from "../util/custom.error";
-import { respond } from "../utils/respond";
-import { assertEnvVar } from "../validation/assert";
+import { respond } from "../util/respond";
+import { assertEnvVar } from "../util/assert";
 import { AwesomeMessagesService } from "./AwesomeMessagesService";
 import { ExpoSendAdapter } from "./ExpoSendAdapter";
 import { Sender } from "./Sender";
